requirements.txt
pytest==5.3.2
pytest-cov==2.8.1
Flask==1.1.1
skia-python==87.5